refactor: marketplace tool folder
